part_4_streamlit/requirements.txt
streamlit
pymongo
pandas
plotly
python-dotenv
statsmodels
